discovery: move TargetSet into discovery package
